# backend/app/models/common.py
from enum import Enum
from pydantic import BaseModel, Field

class SupportedLanguage(str, Enum):
    JAVASCRIPT = "javascript"
    TYPESCRIPT = "typescript"
    PYTHON = "python"
    JAVA = "java"
    CPP = "cpp"
    GO = "go"

class Role(str, Enum):
    HOST = "host"
    PARTICIPANT = "participant"
    SPECTATOR = "spectator"

class CursorPosition(BaseModel):
    line: int = Field(ge=1, le=100_000)
    column: int = Field(ge=1, le=100_000)