# backend/openapi.yaml
{
  "components": {
    "schemas": {
...
      "Difficulty": {
        "enum": [
          "easy",
          "medium",
          "hard"
        ],
        "title": "Difficulty",
        "type": "string"
      },
      "Example": {
        "properties": {
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "input": {
            "title": "Input",
            "type": "string"
          },
          "output": {
            "title": "Output",
            "type": "string"
          }
        },
        "required": [
          "input",
          "output"
        ],
        "title": "Example",
        "type": "object"
      },
      "ExecutionRequest": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/SupportedLanguage"
          }
        },
        "required": [
          "code",
          "language"
        ],
        "title": "ExecutionRequest",
        "type": "object"
      },
      "ExecutionResult": {
        "properties": {
          "executionTime": {
            "title": "Executiontime",
            "type": "number"
          },
          "exitCode": {
            "title": "Exitcode",
            "type": "integer"
          },
          "stderr": {
            "title": "Stderr",
            "type": "string"
          },
          "stdout": {
            "title": "Stdout",
            "type": "string"
          },
          "testResults": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TestResult"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Testresults"
          }
        },
        "required": [
          "stdout",
          "stderr",
          "exitCode",
          "executionTime"
        ],
        "title": "ExecutionResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Problem": {
        "properties": {
          "constraints": {
            "items": {
              "type": "string"
            },
            "title": "Constraints",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "examples": {
            "items": {
              "$ref": "#/components/schemas/Example"
            },
            "title": "Examples",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "examples",
          "constraints",
          "difficulty"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "Role": {
        "enum": [
          "host",
          "participant",
          "spectator"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "SupportedLanguage": {
        "enum": [
          "javascript",
          "typescript",
          "python",
          "java",
          "cpp",
          "go"
        ],
        "title": "SupportedLanguage",
        "type": "string"
      },
      "TestRequest": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "language": {
            "$ref": "#/components/schemas/SupportedLanguage"
          },
          "problem": {
            "$ref": "#/components/schemas/Problem"
          }
        },
        "required": [
          "code",
          "language",
          "problem"
        ],
        "title": "TestRequest",
        "type": "object"
      },
      "TestResult": {
        "properties": {
          "actual": {
            "title": "Actual",
            "type": "string"
          },
          "expected": {
            "title": "Expected",
            "type": "string"
          },
          "input": {
            "title": "Input",
            "type": "string"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          }
        },
        "required": [
          "passed",
          "input",
          "expected",
          "actual"
        ],
        "title": "TestResult",
        "type": "object"
      },
      "User": {
        "properties": {
          "avatar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar"
          },
          "createdAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ],
        "title": "User",
        "type": "object"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "password": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "username": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserLogin": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "session_token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend API for CodioLive (Mock Interview Hub)",
    "title": "CodioLive API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Login with email and password.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Revoke and clear the current session cookie.",
        "operationId": "logout_auth_logout_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "session_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "description": "Get the current authenticated user.",
        "operationId": "get_current_user_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Current User",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "description": "Create a new account.",
        "operationId": "signup_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup",
        "tags": [
          "Auth"
        ]
      }
    },
    "/execution/run": {
      "post": {
        "description": "Code execution is disabled on the server for security reasons.\nAn isolated execution service is required before code can run.",
        "operationId": "run_code_execution_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Code",
        "tags": [
          "Execution"
        ]
      }
    },
    "/execution/test": {
      "post": {
        "description": "Test execution is disabled on the server for security reasons.\nAn isolated execution service is required before tests can run.",
        "operationId": "run_tests_execution_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Tests",
        "tags": [
          "Execution"
        ]
      }
    },
...
  }
}